pyproject.toml
[project]
name = "brontes"
version = "0.1.0"
description = "Agent application using Strands Agents"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "strands-agents[otel]>=0.1.0",
    "strands-agents-tools>=0.2.0",
    "nbformat>=5.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []
